rename photoQuestion to havePhoto
