+++ integration_tests/pages/components/forms/attachments/HavePhotoForm.ts
@@ -1,8 +1,8 @@
 import FormComponent from '../../formComponent'
 import FormRadiosComponent from '../../formRadiosComponent'
 
-export default class PhotoQuestionFormComponent extends FormComponent {
-  get photoQuestionField(): FormRadiosComponent {
+export default class HavePhotoFormComponent extends FormComponent {
+  get havePhotoField(): FormRadiosComponent {
     const label = 'Do you have a photo to upload?'
     return new FormRadiosComponent(this.form, label, ['Yes', 'No'])
   }
