+++ integration_tests/scenarios/SR01 Adult Device Wearer/Trail monitoring/CEMO021 -  Immigration - GPS tag location fitted - variation of additional address.cy.ts
@@ -23,7 +23,7 @@ import ContactInformationCheckYourAnswersPage from '../../../pages/order/contact
 import IdentityNumbersPage from '../../../pages/order/about-the-device-wearer/identity-numbers'
 import InstallationLocationPage from '../../../pages/order/monitoring-conditions/installation-location'
 import UploadLicencePage from '../../../pages/order/attachments/uploadLicence'
-import PhotoQuestionPage from '../../../pages/order/attachments/photoQuestion'
+import HavePhotoPage from '../../../pages/order/attachments/havePhoto'
 
 context.skip('Scenarios', () => {
   const fmsCaseId: string = uuidv4()
@@ -206,8 +206,8 @@ context.skip('Scenarios', () => {
       licencePage.form.uploadField.uploadFile({ fileName: files.licence.fileName, contents: files.licence.contents })
       licencePage.form.saveAndContinueButton.click()
 
-      const havePhotoPage = Page.verifyOnPage(PhotoQuestionPage)
-      havePhotoPage.form.photoQuestionField.set('No')
+      const havePhotoPage = Page.verifyOnPage(HavePhotoPage)
+      havePhotoPage.form.havePhotoField.set('No')
       havePhotoPage.form.saveAndContinueButton.click()
 
       const attachmentPage = Page.verifyOnPage(AttachmentSummaryPage)
