+++ integration_tests/scenarios/SR02 Child Device Wearer/Attendance monitoring/CEMO035 - Community - Youth Rehabilitation Order with GPS Tag.cy.ts
@@ -29,7 +29,7 @@ import IdentityNumbersPage from '../../../pages/order/about-the-device-wearer/id
 import AttendanceMonitoringPage from '../../../pages/order/monitoring-conditions/attendance-monitoring'
 import InstallationLocationPage from '../../../pages/order/monitoring-conditions/installation-location'
 import UploadLicencePage from '../../../pages/order/attachments/uploadLicence'
-import PhotoQuestionPage from '../../../pages/order/attachments/photoQuestion'
+import HavePhotoPage from '../../../pages/order/attachments/havePhoto'
 
 // test disabled as community YRO is not currently a valid sentence type
 context.skip('Scenarios', () => {
@@ -239,8 +239,8 @@ context.skip('Scenarios', () => {
         licencePage.form.uploadField.uploadFile({ fileName: files.licence.fileName, contents: files.licence.contents })
         licencePage.form.saveAndContinueButton.click()
 
-        const havePhotoPage = Page.verifyOnPage(PhotoQuestionPage)
-        havePhotoPage.form.photoQuestionField.set('No')
+        const havePhotoPage = Page.verifyOnPage(HavePhotoPage)
+        havePhotoPage.form.havePhotoField.set('No')
         havePhotoPage.form.saveAndContinueButton.click()
 
         const attachmentPage = Page.verifyOnPage(AttachmentSummaryPage)
