+++ integration_tests/scenarios/SR02 Child Device Wearer/Exclusion zone/CEMO014 - post release - multiple exclusion zones - single document attachment.cy.ts
@@ -29,7 +29,7 @@ import IdentityNumbersPage from '../../../pages/order/about-the-device-wearer/id
 import InstallationAndRiskCheckYourAnswersPage from '../../../pages/order/installation-and-risk/check-your-answers'
 import InstallationLocationPage from '../../../pages/order/monitoring-conditions/installation-location'
 import UploadLicencePage from '../../../pages/order/attachments/uploadLicence'
-import PhotoQuestionPage from '../../../pages/order/attachments/photoQuestion'
+import HavePhotoPage from '../../../pages/order/attachments/havePhoto'
 
 context('Scenarios', () => {
   const fmsCaseId: string = uuidv4()
@@ -228,8 +228,8 @@ context('Scenarios', () => {
         licencePage.form.uploadField.uploadFile({ fileName: uploadFile.fileName, contents: uploadFile.contents })
         licencePage.form.saveAndContinueButton.click()
 
-        const havePhotoPage = Page.verifyOnPage(PhotoQuestionPage)
-        havePhotoPage.form.photoQuestionField.set('No')
+        const havePhotoPage = Page.verifyOnPage(HavePhotoPage)
+        havePhotoPage.form.havePhotoField.set('No')
         havePhotoPage.form.saveAndContinueButton.click()
 
         const attachmentPage = Page.verifyOnPage(AttachmentSummaryPage)
