+++ integration_tests/scenarios/kitchen-sink.cy.ts
@@ -33,7 +33,7 @@ import ProbationDeliveryUnitPage from '../pages/order/contact-information/probat
 import CurfewAdditionalDetailsPage from '../pages/order/monitoring-conditions/curfew-additional-details'
 import InstallationLocationPage from '../pages/order/monitoring-conditions/installation-location'
 import UploadLicencePage from '../pages/order/attachments/uploadLicence'
-import PhotoQuestionPage from '../pages/order/attachments/photoQuestion'
+import HavePhotoPage from '../pages/order/attachments/havePhoto'
 
 context('The kitchen sink', () => {
   const takeScreenshots = config.screenshots_enabled
@@ -403,9 +403,9 @@ context('The kitchen sink', () => {
       if (takeScreenshots) cy.screenshot('22. licencePage', { overwrite: true })
       licencePage.form.saveAndContinueButton.click()
 
-      const havePhotoPage = Page.verifyOnPage(PhotoQuestionPage)
+      const havePhotoPage = Page.verifyOnPage(HavePhotoPage)
       if (takeScreenshots) cy.screenshot('23. havePhotoPage', { overwrite: true })
-      havePhotoPage.form.photoQuestionField.set('No')
+      havePhotoPage.form.havePhotoField.set('No')
       havePhotoPage.form.saveAndContinueButton.click()
 
       const attachmentPage = Page.verifyOnPage(AttachmentSummaryPage)
