+++ integration_tests/scenarios/mandatory-field-only.cy.ts
@@ -38,7 +38,7 @@ import ProbationDeliveryUnitPage from '../pages/order/contact-information/probat
 import CurfewAdditionalDetailsPage from '../pages/order/monitoring-conditions/curfew-additional-details'
 import InstallationLocationPage from '../pages/order/monitoring-conditions/installation-location'
 import UploadLicencePage from '../pages/order/attachments/uploadLicence'
-import PhotoQuestionPage from '../pages/order/attachments/photoQuestion'
+import HavePhotoPage from '../pages/order/attachments/havePhoto'
 
 context('Mandatory fields only', () => {
   const takeScreenshots = config.screenshots_enabled
@@ -374,9 +374,9 @@ context('Mandatory fields only', () => {
       if (takeScreenshots) cy.screenshot('22. licencePage', { overwrite: true })
       licencePage.form.saveAndContinueButton.click()
 
-      const havePhotoPage = Page.verifyOnPage(PhotoQuestionPage)
+      const havePhotoPage = Page.verifyOnPage(HavePhotoPage)
       if (takeScreenshots) cy.screenshot('23. havePhotoPage', { overwrite: true })
-      havePhotoPage.form.photoQuestionField.set('No')
+      havePhotoPage.form.havePhotoField.set('No')
       havePhotoPage.form.saveAndContinueButton.click()
 
       const attachmentPage = Page.verifyOnPage(AttachmentSummaryPage)
@@ -684,9 +684,9 @@ context('Mandatory fields only', () => {
       if (takeScreenshots) cy.screenshot('22. licencePage', { overwrite: true })
       licencePage.form.saveAndContinueButton.click()
 
-      const havePhotoPage = Page.verifyOnPage(PhotoQuestionPage)
+      const havePhotoPage = Page.verifyOnPage(HavePhotoPage)
       if (takeScreenshots) cy.screenshot('23. havePhotoPage', { overwrite: true })
-      havePhotoPage.form.photoQuestionField.set('No')
+      havePhotoPage.form.havePhotoField.set('No')
       havePhotoPage.form.saveAndContinueButton.click()
 
       const attachmentPage = Page.verifyOnPage(AttachmentSummaryPage)
