+++ server/controllers/attachments/attachmentHavePhotoController.ts
