+++ server/models/view-models/additionalDocumentsCheckAnswers.ts
@@ -18,7 +18,7 @@ const createViewModel = (order: Order, content: I18n | undefined) => {
       order,
     ),
     createAnswer(
-      content?.pages.photoQuestion.questions.havePhoto.text || '',
+      content?.pages.havePhoto.questions.havePhoto.text || '',
       order.orderParameters?.havePhoto ? 'Yes' : 'No',
       paths.ATTACHMENT.PHOTO_QUESTION.replace(':orderId', order.id),
       answerOpts,
