+++ server/types/i18n/index.ts
@@ -21,7 +21,7 @@ import VariationDetailsPageContent from './pages/variationDetails'
 import InstallationLocationPageContent from './pages/installationLocation'
 import InstallationAppointmentPageContent from './pages/installationAppointment'
 import ReferenceData from './reference'
-import PhotoQuestionPageContent from './pages/photoQuestion'
+import havePhotoPageContent from './pages/havePhoto'
 
 type I18n = {
   pages: {
@@ -48,7 +48,7 @@ type I18n = {
     trailMonitoring: TrailMonitoringPageContent
     uploadLicense: UploadDocumentPageContent
     uploadPhotoId: UploadDocumentPageContent
-    photoQuestion: PhotoQuestionPageContent
+    havePhoto: havePhotoPageContent
     variationDetails: VariationDetailsPageContent
     installationLocation: InstallationLocationPageContent
     installationAppointment: InstallationAppointmentPageContent
