+++ server/types/i18n/pages/havePhoto.ts
@@ -1,0 +1,5 @@
+import PageContent from './page'
+
+type havePhotoPageContent = PageContent<'havePhoto'>
+
+export default havePhotoPageContent
