--- server/types/i18n/pages/photoQuestion.ts
@@ -1,5 +1,0 @@
-import PageContent from './page'
-
-type PhotoQuestionPageContent = PageContent<'havePhoto'>
-
-export default PhotoQuestionPageContent
