+++ server/views/pages/order/attachments/have-photo.njk
@@ -1,8 +1,8 @@
 {% extends "../../../partials/form-layout.njk" %}
 {% from "govuk/components/radios/macro.njk" import govukRadios %}
 
-{% set section = content.pages.photoQuestion.section %}
-{% set questions = content.pages.photoQuestion.questions %}
+{% set section = content.pages.havePhoto.section %}
+{% set questions = content.pages.havePhoto.questions %}
 
 {% block formInputs %}
